Added the ability to print a server's index by supplying "list" as the site argument
Also added the ability to print the server's router version by supplying "version" as the site argument. The site argument should be renamed now, since it's holding more stuff than just the site. And documentation needs to be updated to reflect that these names can't be used in the index file. Better still, have the commands be prefixed with a double dash. Yeah.