Added the ability to define an appropriate user to which to change
The client has been completely rewritten in Python. This is good, because Python makes it much easier to manage the now expanded set of argument flags. But it badly needs commenting.

The router has been expanded to handle its new responsibilities.

The syntax of the index has been expanded to include the user to which to switch.

Both the client and the router could use some refactoring before further significant additions. Mostly in the realm of more intuitive ordering of tasks, and of variable naming and use.